pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "Backend side for autobid24"
authors = ["Martynas Jakimcikas <martynas.jakimcikas@vinted.com>", "Tadas Audinis <t.audinis@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.13"
uvicorn = "^0.34.3"
sqlalchemy = "^2.0.41"
psycopg2-binary = "^2.9.10"
asyncpg = "^0.30.0"
pydantic-settings = "^2.9.1"
databases = "^0.9.0"
greenlet = "^3.2.3"
httpx = "^0.28.1"
dependency-injector = "^4.48.1"
coloredlogs = "^15.0.1"
alembic = "^1.16.2"
isort = "^6.0.1"
flake8 = "^7.3.0"
ruff = "^0.12.1"
pre-commit = "^4.2.0"
black = "^25.1.0"
apscheduler = "^3.11.0"
factory-boy = "^3.3.3"
authlib = "^1.6.0"
beautifulsoup4 = "^4.13.4"
pyotp = "^2.9.0"
cachetools = "^6.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^1.0.0"

[tool.black]
line-length = 120
skip-string-normalization = true
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3

[tool.ruff]
line-length = 120
target-version = "py312"
lint.select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
lint.ignore = [
    "E203",  # whitespace before ':'
    "W291",  # trailing whitespace
    "W292",  # no newline at end of file
]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    ".pytest_cache",
    "migrations"
]

[tool.ruff.lint.isort]
force-single-line = false
force-sort-within-sections = false
known-first-party = []

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

